Collapsed all key point functions into ControlHandle class
Rather than representing the center of the key frame point with a button
and the control handles with a separate object, they were rolled into
one in the ControlHandle class. The grid object is passed to the
ControlHandle class, so most of the position updating is handled within
the class, rather than externally, has had been the case.